Split useBookmarkWeaver hook responsibilities
